Further implementation of UpdateProfile Screen
Further implementation of UpdateProfile screen. Currently it loads in the user profile picture correctly but there is a bug in loading organization and username. changing the picture works but doesn't update in the conversation screen unless manually refreshed. Changing username and org  also seems to be bugged. Also small changes to ProfileSetup.JS and Globals.js that don't effect functionality outside of ProfileUpdate